requirements.txt
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email-validator==1.1.3
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-mongoengine==1.0.0
Flask-WTF==1.2.1
Flask-Limiter==2.4.5
gunicorn==20.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.24.1
packaging==24.1
Pillow==8.4.0
pymongo==3.11.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
setuptools==70.2.0
typing_extensions==4.12.2
Werkzeug>=2.3.0,<3.0
WTForms==3.1.2
PyJWT==2.8.0
cryptography>=41.0.0
honcho==1.1.0
